+++ packages/validasi_annotation/lib/src/key.dart
@@ -1,0 +1,3 @@
+abstract class ValidasiKey<T> {
+  const ValidasiKey();
+}