+++ packages/validasi_annotation/lib/validasi_annotation.dart
@@ -1,5 +1,6 @@
 export 'src/annotation.dart';
 export 'src/base.dart';
+export 'src/key.dart';
 export 'src/rules/common/required.dart';
 export 'src/rules/common/nullable.dart';
 export 'src/rules/min_length.dart';