+++ packages/validasi_gen/lib/src/builder.dart
@@ -2,10 +2,15 @@ import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';
 
 import 'package:validasi_gen/src/generator.dart';
+import 'package:validasi_gen/src/utils.dart';
 
-Builder validasiBuilder(BuilderOptions options) => _ValidasiBuilder();
+Builder validasiBuilder(BuilderOptions options) => _ValidasiBuilder(options);
 
 class _ValidasiBuilder implements Builder {
+  _ValidasiBuilder(this.options);
+
+  final BuilderOptions options;
+
   @override
   Map<String, List<String>> get buildExtensions => const {
         '.dart': ['.g.dart'],
@@ -14,7 +19,11 @@ class _ValidasiBuilder implements Builder {
   @override
   Future<void> build(BuildStep buildStep) async {
     final library = await buildStep.inputLibrary;
-    final output = ValidasiGenerator().generate(
+    final generateFieldsDefault =
+        boolOption(options.config, 'generateFields') ?? true;
+    final output = ValidasiGenerator(
+      generateFieldsDefault: generateFieldsDefault,
+    ).generate(
       LibraryReader(library),
       buildStep,
     );